Draw in Playing Field
Added in code to draw in the playing field. Weird issue with ScreenWidth - the tutorial I'm following had it at 80, but it doesn't draw to the screen correctly unless I set it to 120? Need to look more into how the screen buffer we are using works in order to fidure this out probably.